+++ src/model/entidades_model.py
@@ -1,0 +1,11 @@
+import spacy
+def reconocer_entidad(frase):
+    nlp = spacy.load('es_core_news_lg')
+    documento = nlp(frase)
+    for entidad in documento.ents:
+        entidad = {
+            entidad.text : entidad.label_
+
+        }
+        return entidad
+
