+++ src/services/entidades_services.py
@@ -1,23 +1,11 @@
-import spacy
+from src.model.entidades_model import reconocer_entidad
 
 class entidadesServices():
     @classmethod
     def entidades_texto(cls,texto):
-        
         resultado = []
-
         for frase in texto:
             entidades= reconocer_entidad(frase)
             resultado.append({"oracion":frase, "entidades": entidades})
         return resultado;
 
-def reconocer_entidad(frase):
-    nlp = spacy.load('es_core_news_sm')
-    documento = nlp(frase)
-    for entidad in documento.ents:
-        entidad = {
-            entidad.text : entidad.label_
-
-        }
-        return entidad
-
